requirements.txt
customtkinter==5.2.1
numpy==1.26.1
Pillow==10.0.0
Pillow==10.1.0
